Crate git_conflict

Git Conflict Manager The library is intended to solve the recurring conflict issues in git The crate offers four main strategies to solve the conflict problem which are:

Accept local: it will accept the local changes and ignore the foreign branch’s changes.

Accept Foreign: it will accept the foreign branch’s changes and ignore the local changes.

combination: it will combine both branch’s changes and remove the markers (straightforward). All of which methodologies can cause a complex, diverged or broken trees in the terminal

Trees in git are essentials as they draw the chronological modification made on a Repository Then the commit is the stamp that provides extra more details such as the user who made the commit, time of when the commit was made and signature. Due to that, another advanced method targeted for experienced developers was introduced which is Tree Merge. Tree Merge merges the tree for either (head, foreign) branches with the past tree or the ancestor. The user get to chose the base tree (head, foreign), he also get to chose whether the interference of the past (parent) should be enabled or not. if parent interference is enabled then the current tree selected by the user is directly merged to the ancestor of the conflicted branches, resulting in a smooth straightforward tree. The method is super useful for expert developers as they don’t need to worry about bucking up a specific branch if a conflict occurs, they simply chose parent interference and the branch will be merged with the changes of direct ancestor.

Modules

combine

discarding

git_src

merge_trees

The merge tree struct it applies the logic of merging conflicted tree with the interference of the past for further explanation of this methodologies refer to this article (Soon)

utils

Define wrapper utils for efficient operation This module defines the utility functions required to ensure the smoothness of the conflict resolution operation, for example the Index and Repository struct in lib git2 crate does not implement the Copy and Clone traits. Those Traits are necessary to share the parameters across different functions. Being able to mutually mutate the Index and Repository enables efficient operation and avoid redundant code. Another major factor is the fact that those structs are extensivly used, so the usage of Reference counter is necessary that provides mutliple versions of one share instance in the memory, when combined with the ref cell, it allows for mutual mutation.

Traits

Actions

A trait that contains the necessary action required by all methodologies to resolve a conflict

GitOps

Trait to simply call the methodologies in a simple straightforward manner

Initialize

The initialization trait for which the repository is built

ManualControl

the manual control trait that inherits the Actions trait defines the a manual commiting action.

Utils

The utils trait is used by the utility module