git_conflict

Trait Actions

pub trait Actions {
    // Required methods
    fn branches(&self) -> Branches;
    fn index(&self) -> RefMut<'_, Index>;
    fn repo(&self) -> RefMut<'_, Repository>;

    // Provided methods
    fn staging(&mut self, files: Vec<String>) { ... }
    fn commit(&mut self, parent_commits: &[Oid], msg: String) -> bool { ... }
    fn does_conflict_exists(&self) -> bool { ... }
    fn return_conflicted_files(&self, condition: Status) -> Option<Vec<String>> { ... }
}

A trait that contains the necessary action required by all methodologies to resolve a conflict

Required Methods

fn branches(&self) -> Branches

fn index(&self) -> RefMut<'_, Index>

fn repo(&self) -> RefMut<'_, Repository>

Provided Methods

fn staging(&mut self, files: Vec<String>)

fn commit(&mut self, parent_commits: &[Oid], msg: String) -> bool

Making a commit this function has an embedding implementation

fn does_conflict_exists(&self) -> bool

checking whether a conflict exists or not

fn return_conflicted_files(&self, condition: Status) -> Option<Vec<String>>

return the file with a specific conditions, eg., Conflicted